+++ docs/index.html
@@ -110,7 +110,7 @@
 					petfinderAPI.init({
 						key: key,
 						shelter: shelter,
-						callbackAfter: function () {
+						callback: function () {
 							petfinderSort.init();
 							petfinderImgToggle.init();
 						}
