+++ src/docs/_templates/_footer.html
@@ -40,7 +40,7 @@
 					petfinderAPI.init({
 						key: key,
 						shelter: shelter,
-						callbackAfter: function () {
+						callback: function () {
 							petfinderSort.init();
 							petfinderImgToggle.init();
 						}
